Fix non-mutex check not accounting for jump offset
Since the unconditional jump instruction uses offsets
and not absolute addresses, we can't just dereference
the pointer to the jump address and expect that to be
the actual hook function's address.

To counter this, add the pointer's address to the
offset found, and add another 5 due to the five bytes
occupied by the jump itself (0xE9 + four byte offset).