Revert all changes adopting the GZCOM framework
There are numerous issues with using it, for example the
fact that there could be multiple SC4Fix DLLs loaded that
all have the same director ID. There's also the fact that
despite using the same patches as before, the game now
crashes upon loading it.

Changes to version checking will still hold. DLL unload
preemption is only staying by virtue of necessity for
SC4Fix itself to stay loaded, and if I could get this
to work without it, I would. It's an ugly and terrible
hack.